# typescript/packages/apps/reference-datasets/docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SIF SaaS: Reference Datasets Management",
    "description": "\nHas accountability for:\n- TODO...\n",
    "version": "0.0.1"
  },
  "components": {
    "securitySchemes": {
      "platformUserPool": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header"
      }
    },
    "schemas": {
      "tags": {
        "description": "User-defined searchable tags",
        "type": "object",
        "additionalProperties": {
          "anyOf": [
            {
              "type": "string"
            },
            {
              "type": "null"
            }
          ]
        }
      },
      "referenceDatasetResource": {
        "type": "object",
        "properties": {
          "createdAt": {
            "description": "Date/time created",
            "format": "date-time",
            "type": "string"
          },
          "createdBy": {
            "description": "ID of owner.",
            "type": "string"
          },
          "datasetHeaders": {
            "type": "array",
            "items": {
              "description": "datasetHeaders form-data attribute is required attribute to pass csv file headers for validation. A thing to note\nis that this attribute should be specified as form-data attribute which also has content-type specified and is set  to 'application/json'. This is not immediately known or visible.\nIts recommended to check how your client can set the content-type to 'application/json' for individual form-data attributes. As an example this is how you can set the content-type\non an individual form-data attribute for a CURL reques '--form 'datasetHeaders=\"[\\\"ZIP\\\", \\\"STATE\\\"]\";type=application/json' \\' ",
              "type": "string"
            }
          },
          "description": {
            "description": "description for the reference dataset",
            "type": "string"
          },
          "groups": {
            "description": "Groups that have access to the calculation.",
            "type": "array",
            "items": {
              "description": "The group ID as managed by the Access Management module.",
              "type": "string"
            }
          },
          "id": {
            "description": "Unique id.",
            "type": "string"
          },
          "name": {
            "description": "name of the ReferenceDataset",
            "type": "string"
          },
          "state": {
            "type": "string",
            "enum": [
              "enabled",
              "frozen",
              "disabled"
            ],
            "description": "State of the resource. `enabled` allows use of this version of the resource within transforms. `disabled` allows this version of the resource to be used within existing transforms but not allowed to be used as part of new transforms. `frozen` prevents any version of the resource from use within any transforms including pre-existing."
          },
          "status": {
            "type": "string",
            "enum": [
              "pendingUpload",
              "inProgress",
              "success",
              "failed"
            ],
            "description": "reference dataset processing status"
          },
          "statusMessage": {
            "description": "message related to the status attribute. e.g. if the status is failed, this attribute will indicate why it failed",
            "type": "string"
          },
          "tags": {
            "$ref": "#/components/schemas/tags"
          },
          "updatedAt": {
            "description": "Date/time updated",
            "format": "date-time",
            "type": "string"
          },
          "updatedBy": {
            "description": "Last ID of user who made a change.",
            "type": "string"
          },
          "uploadUrl": {
            "description": "signed url for uploading the reference dataset",
            "type": "string"
          },
          "activeAt": {
            "description": "Date/time resource is active",
            "format": "date-time",
            "type": "string"
          },
          "version": {
            "description": "version of the reference dataset",
            "type": "number"
          }
        },
        "required": [
          "createdAt",
          "createdBy",
          "groups",
          "id",
          "name",
          "state",
          "status",
          "version"
        ]
      },
...
      "referenceDatasetVersionList": {
        "type": "object",
        "properties": {
          "referenceDatasets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/referenceDatasetResource"
            }
          },
          "pagination": {
            "type": "object",
            "properties": {
              "lastEvaluatedVersion": {
                "description": "version of the reference dataset",
                "type": "number"
              }
            }
          }
        },
        "required": [
          "referenceDatasets"
        ]
      },
...
    }
  },
  "paths": {
...
    "/referenceDatasets/{id}/versions": {
      "get": {
        "operationId": "listVersions",
        "tags": [
          "Reference Datasets"
        ],
        "description": "Lists Reference Datasets by versions\n\nPermissions:\n- All roles of the group in context may lists Reference Datasets by versions.\n",
        "parameters": [
          {
            "schema": {
              "type": "integer"
            },
            "in": "query",
            "name": "count",
            "required": false,
            "description": "Count of results to return."
          },
          {
            "schema": {
              "type": "integer"
            },
            "in": "query",
            "name": "fromVersion",
            "required": false,
            "description": "Version to paginate from (exclusive)."
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "versionAsAt",
            "required": false,
            "description": "Filters by `asAt` or `createdAt` determine the latest resource version as that time."
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Unique id."
          },
          {
            "in": "header",
            "name": "accept-version",
            "required": true,
            "description": "API version",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "accept",
            "required": true,
            "description": "Accepted Content Type",
            "schema": {
              "type": "string"
            }
          }
        ],
        "x-security-scopes": [
          "reader",
          "contributor",
          "admin",
          "superAdmin"
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Success.",
                  "$ref": "#/components/schemas/referenceDatasetVersionList"
                },
                "examples": {
                  "List of reference datasets by version": {
                    "summary": "Paginated list of reference datasets by version",
                    "value": {
                      "pagination": {
                        "lastEvaluatedVersion": 2
                      },
                      "referenceDatasets": [
                        {
                          "id": "defa5ad0-2938-11ed-b188-3789cc690104",
                          "createdAt": "2022-08-31T14:26:15.421Z",
                          "createdBy": "someone@example.com",
                          "name": "dataset one",
                          "groups": [
                            "/usa/northwest"
                          ],
                          "version": 1,
                          "datasetHeaders": [
                            "Type",
                            "Units"
                          ],
                          "tags": {
                            "division": "purchasing",
                            "type": "material/metal/steel"
                          },
                          "state": "frozen",
                          "status": "success"
                        },
                        {
                          "id": "fd06ba10-2937-11ed-ae7b-0579fa02d3cb",
                          "createdAt": "2022-08-31T14:26:15.421Z",
                          "createdBy": "someone@example.com",
                          "updatedBy": "someone@example.com",
                          "updatedAt": "2022-08-31T14:19:56.337Z",
                          "name": "dataset one",
                          "groups": [
                            "/usa/northwest"
                          ],
                          "version": 2,
                          "datasetHeaders": [
                            "Type",
                            "Measurement"
                          ],
                          "tags": {
                            "division": "purchasing",
                            "type": "material/metal/steel"
                          },
                          "state": "enabled",
                          "status": "success"
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Forbidden.",
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/tags/{key}": {
      "get": {
        "operationId": "listTags",
        "summary": "List tag values",
        "tags": [
          "Tags"
        ],
        "description": "List tags for a given key.\n\nFor hierarchical / cascading tags, provide the `?parentValue` query string parameter.\n\nPermissions:\n- `reader` and above may list tags.\n",
        "parameters": [
          {
            "schema": {
              "type": "integer"
            },
            "in": "query",
            "name": "count",
            "required": false,
            "description": "Count of results to return."
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "fromValue",
            "required": false,
            "description": "Tag value to paginate from (exclusive)."
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "parentValue",
            "required": false,
            "description": "Chosen parent value for use with hierarchical tags."
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "key",
            "required": true,
            "description": "Tag key."
          },
          {
            "in": "header",
            "name": "accept-version",
            "required": true,
            "description": "API version",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "accept",
            "required": true,
            "description": "Accepted Content Type",
            "schema": {
              "type": "string"
            }
          }
        ],
        "x-security-scopes": [
          "reader",
          "contributor",
          "admin",
          "superAdmin"
        ],
        "responses": {
          "200": {
            "description": "List of tag values",
            "content": {
              "application/json": {
                "schema": {
                  "description": "List of tag values",
                  "type": "object",
                  "properties": {
                    "values": {
                      "description": "Tag values and labels.",
                      "type": "object",
                      "additionalProperties": {
                        "type": "string"
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "count": {
                          "description": "No. of results returned when pagination requested.",
                          "type": "integer"
                        },
                        "lastEvaluatedValue": {
                          "description": "Last evaluated tag value",
                          "type": "string"
                        }
                      }
                    }
                  },
                  "required": [
                    "values"
                  ]
                },
                "examples": {
                  "List of tag values": {
                    "summary": "Paginated list of tag values.",
                    "value": {
                      "values": [
                        {
                          "ghg protocol": "GHG Protocol",
                          "the green alliance": "The Green Alliance"
                        }
                      ],
                      "pagination": {
                        "count": 2,
                        "lastEvaluated": {
                          "value": "the green alliance"
                        }
                      }
                    },
                    "List of hierarchical tag values": {
                      "summary": "Paginated list of hierarchical tag values.",
                      "value": {
                        "values": [
                          {
                            "material/metal/steel": "Steel",
                            "material/metal/iron": "Iron"
                          }
                        ],
                        "pagination": {
                          "count": 2,
                          "lastEvaluated": {
                            "value": "material/metal/iron"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Forbidden.",
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                }
              }
            }
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost"
    }
  ],
  "security": [],
...
}